fix(scan): strengthen hook guard test, extract config methods to app_config.go
